Complete build fixes and add modernization features - working application

+++ routes/budgets.go
@@ -18,10 +18,10 @@ func RegisterBudgetRoutes(r *mux.Router) {
 }
 
 type BudgetRequest struct {
-	Name     string `json:"name"`
-	Category string `json:"category"`
-	Amount   int64  `json:"amount"`
-	Period   string `json:"period"`
+	Name     string  `json:"name"`
+	Category string  `json:"category"`
+	Amount   float64 `json:"amount"`
+	Period   string  `json:"period"`
 }
 
 func getBudgets(w http.ResponseWriter, r *http.Request) {
